requirements.txt
arrow==1.2.1
opencv-python==4.5.3.56
Pillow==8.3.2
PyAutoGUI==0.9.53
